[REFACTOR/#39] LottieAnimation core:ui로 이동 및 scale 비율 계산 헬퍼 함수로 빼기
