+++ core/ui/build.gradle.kts
@@ -14,4 +14,5 @@ dependencies {
     implementation(libs.orbit.core)
     implementation(libs.orbit.compose)
     implementation(libs.orbit.viewmodel)
+    implementation(libs.lottie.compose)
 }
