+++ feature/onboarding/build.gradle.kts
@@ -13,7 +13,6 @@ dependencies {
     implementation(libs.orbit.core)
     implementation(libs.orbit.compose)
     implementation(libs.orbit.viewmodel)
-    implementation(libs.lottie.compose)
     implementation(libs.coil.compose)
     implementation(libs.coil.gif)
 }
