+++ feature/onboarding/src/main/java/com/kms/onboarding/OnboardingCompleteScreen1.kt
@@ -12,8 +12,8 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import com.kms.onboarding.component.LottieAnimation
 import com.yapp.designsystem.theme.OrbitTheme
+import com.yapp.ui.component.lottie.LottieAnimation
 import com.yapp.ui.utils.heightForScreenPercentage
 import com.yapp.ui.utils.paddingForScreenPercentage
 import feature.onboarding.R
